Excel.Addin/UI: add domain override option
Update settings to include optional domain override text box.  If users
want to change the endpoint to another location they can do so with this
option.

Updates the Visual Studio designer xml to include the new label, text
box and reset button.  While we're at it, adding a new grid row allows
us to update the appropriate Grid.Row indices as they are currently
incorrect.  Update the offsets accordingly.  We now have 7 grid rows.

The reset button to allow users to unset a domain to switch back to the
default.

A future commit will handle validating user input and saving to the
registry.

Signed-off-by: Jamie Couture <jamie@quandl.com>