rebrand updates
Excel plugin rebrand changes require asset changes, as well as updates
to documentation URLs and email.

Despite logo_sm.png having a transparent background / alpha channel, it
was necessary to set the background colour to 'white' to help mask the
logo.  Not setting Resources.resx to use a white background for the 10th
imageList property would result in a 'white' background on the menu
item, and would look weird.

Refer to addin-product as "Nasdaq Data Link".

Signed-off-by: Jamie Couture <jamie@quandl.com>