Shared.modules: update register path
SetupHelp is not used by main runtime; The execution path begins in
Quandl.Excel.Console/Progam.cs caller to SetupHelp.RegisterExcelAddin().

The derived path should be updated to where the MSI actually installed
the xll file.

The MSI installer will register the UDF library to AppData, not Program
Files x86.  Expand to %AppDat% [1] instead.

Fixes an issue where the wrong path to the UDF xll is set to the
Office/<version>/Excel/Options/OPEN registry value.

While we're at it, fix the source file we're trying to register as the
expected install path assumes the wrong path and name of the file.  One
side effect is that UniversalFunctions is completely broken unless MSI
installer is run first.  The unit tests are completely broken anyway.

Signed-off-by: Jamie Couture <jamie@quandl.com>

--

[1] https://docs.microsoft.com/en-us/windows/win32/shell/knownfolderid?redirectedfrom=MSDN